sysdb: When handling a reply, base further actions on the response code.
This will actually enable the command handler to see whether a log message or
a query response was received.

Also, this ensures that the command handler may correctly detect errors during
query execution and exit with a non-zero code accordingly. The simple_query
integration test was updated to reflect this new, corrected behavior. Thanks
to the test for actually catching this misbehavior! :-)