store_lookup: Added a matcher matching by object name.
This is a generic matcher applied to hosts (or their services or attributes)
and matches based on the respective object name.

Subsequently, this will replace the more complex host/service/attribute
matchers. Instead, more basic matchers will be available which can then be
combined arbitrarily using the existing logical matchers.